Tests/LibWeb/Text/expected/WebAnimations/animation-properties/timeline.txt

Animation's default timeline is the document's timeline: true
Animation created with null timeline has no timeline: true
